SWARM-778 - Usage of our @DefaultDeployment.
Motivation
----------

Since we have @DefaultDeployment, we should actually use it
for our own tests.

Modifications
-------------

Adjusted the Camel-CDI testsuite to use @DefaultDeployment.

Result
------

Simpler code!
