+++ arquillian/arquillian/src/main/java/org/wildfly/swarm/arquillian/deployment/TestableArchiveService.java
@@ -1,7 +1,8 @@
 package org.wildfly.swarm.arquillian.deployment;
 
+import javax.enterprise.inject.Vetoed;
+
 import org.jboss.as.server.deployment.DeploymentUnit;
-import org.jboss.msc.inject.Injector;
 import org.jboss.msc.service.Service;
 import org.jboss.msc.service.ServiceName;
 import org.jboss.msc.service.StartContext;
@@ -13,6 +14,7 @@ import org.wildfly.swarm.arquillian.daemon.server.Server;
 /**
  * @author Bob McWhirter
  */
+@Vetoed
 public class TestableArchiveService implements Service<Void> {
 
     public static final ServiceName NAME = ServiceName.of( "wildfly", "swarm", "arquillian", "testable-archive" );
