+++ testsuite/testsuite-camel-cdi/src/main/java/org/wildfly/swarm/camel/test/cdi/subA/RoutesContextC.java
@@ -7,9 +7,9 @@
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -48,22 +48,22 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /**
- * Uses contextD implicitly using that context for all injection points without
+ * Uses contextC implicitly using that context for all injection points without
  * having to mention them on each camel annotation
  */
-@ContextName("contextD")
-public class RoutesContextD extends RouteBuilder {
-    private static final Logger LOG = LoggerFactory.getLogger(RoutesContextD.class);
+@ContextName("contextC")
+public class RoutesContextC extends RouteBuilder {
+    private static final Logger LOG = LoggerFactory.getLogger(RoutesContextC.class);
 
-    @EndpointInject(uri = "mock:D.b", context = "contextD")
+    @EndpointInject(uri = "mock:C.b", context = "contextC")
     public MockEndpoint b;
 
-    @ContextName("contextD")
-    @Inject @Uri("seda:D.a")
+    @ContextName("contextC")
+    @Inject @Uri("seda:C.a")
     Endpoint a;
 
-    @ContextName("contextD")
-    @Inject @Uri("seda:D.a")
+    @ContextName("contextC")
+    @Inject @Uri("seda:C.a")
     ProducerTemplate producer;
 
     @Override
@@ -72,7 +72,7 @@ public class RoutesContextD extends RouteBuilder {
     }
 
     public void sendMessages() {
-        for (Object expectedBody : Constants.EXPECTED_BODIES_D) {
+        for (Object expectedBody : Constants.EXPECTED_BODIES_C) {
             LOG.info("Sending " + expectedBody + " to " + producer.getDefaultEndpoint());
             producer.sendBody(expectedBody);
         }
