+++ testsuite/testsuite-camel-cxf/src/main/java/org/wildfly/swarm/camel/test/cxf/ws/subA/Endpoint.java
