+++ testsuite/testsuite-camel-cxf/src/main/java/org/wildfly/swarm/camel/test/jaxrs/subA/GreetingServiceImpl.java
