+++ testsuite/testsuite-cdi/src/main/java/org/wildfly/swarm/cdi/ConfigAwareBean.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.wildfly.swarm.cdi.test;
+package org.wildfly.swarm.cdi;
 
 import javax.enterprise.context.Dependent;
 import javax.inject.Inject;
