+++ testsuite/testsuite-cdi/src/test/java/org/wildfly/swarm/cdi/CDIArquillianTest.java
@@ -13,34 +13,36 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.wildfly.swarm.cdi.test;
+package org.wildfly.swarm.cdi;
 
-import java.util.Optional;
-
+import javax.enterprise.inject.spi.CDI;
 import javax.inject.Inject;
 
 import org.jboss.arquillian.junit.Arquillian;
-import org.junit.Assert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.wildfly.swarm.arquillian.DefaultDeployment;
-import org.wildfly.swarm.spi.runtime.annotations.ConfigurationValue;
+import org.wildfly.swarm.cdi.Cheddar;
+
+import static org.junit.Assert.assertNotNull;
 
 /**
- * @author George Gastaldi
+ * @author Bob McWhirter
  */
 @RunWith(Arquillian.class)
 @DefaultDeployment
-public class ConfigurationValueProducerTest {
+public class CDIArquillianTest {
 
     @Inject
-    @ConfigurationValue("logger.level")
-    private Optional<String> loggerLevel;
+    private Cheddar cheddar;
 
     @Test
-    public void testServerAddressExists() {
-        Assert.assertNotNull(loggerLevel);
-        Assert.assertEquals("DEBUG",loggerLevel.get());
+    public void testInjection() {
+        assertNotNull(cheddar);
     }
 
+    @Test
+    public void testCDIContainerPresence() throws Exception {
+        assertNotNull(CDI.current());
+    }
 }
